src/tools/benchmarking/azure/requirements.txt
azure-identity
azure-mgmt-compute
azure-mgmt-network
azure-mgmt-resource
paramiko
python-dotenv
matplotlib
numpy
itertools